+++ pyne/errors.py
@@ -1,0 +1,6 @@
+class NoSouchItemError(Exception):
+    pass
+
+
+class NoSouchPositionError(Exception):
+    pass