+++ pyne/__init__.py
@@ -6,6 +6,7 @@ from . import beatle
 from . import sound
 from .game_gui_controller import GameGUIController
 from . import errors
+from . import constants
 
 __version__ = '0.2.1 alpha'