+++ pyne/__init__.py
@@ -8,6 +8,6 @@ from .game_gui_controller import GameGUIController
 from . import errors
 from . import constants
 
-__version__ = '0.2.1 alpha'
+__version__ = '0.2.1'
 
 print(f'\n[Pyne] version -> {__version__}')